pyproject.toml
[project]
name = "email-composer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.8",
    "google-genai>=1.63.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "rich>=14.3.2",
    "uvicorn>=0.40.0",
]
